+++ packages/orq-rc/src/orq_ai_sdk/models/runagentop.py
@@ -2058,7 +2058,7 @@ class RunAgentAgentToolInputRunTools(BaseModel):
 
     schema_: Annotated[AgentToolInputRunSchema, pydantic.Field(alias="schema")]
 
-    id: Optional[str] = "01M1XDKT6YRZKA91V0K28Z5EQB"
+    id: Optional[str] = "01M1XYQRES5NPNNQ4S2ACYB8Y7"
 
     description: Optional[str] = None