+++ packages/orq-rc/src/orq_ai_sdk/models/updatetoolop.py
@@ -1107,7 +1107,7 @@ class UpdateToolResponseBodyCodeExecutionTool(BaseModel):
     code_tool: UpdateToolResponseBodyCodeTool
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01M1XDKT9H5D1W0JHJ0RF3NPG1"
+        "tool_01M1XYQRHQNN4V9ZMCZBRMQK60"
     )
 
     display_name: Optional[str] = None
@@ -1412,7 +1412,7 @@ class UpdateToolResponseBodyHTTPTool(BaseModel):
     http: UpdateToolResponseBodyHTTP
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01M1XDKT9G607ARACQ4D2EZNC5"
+        "tool_01M1XYQRHPNB4AE6ZKSP25MG5A"
     )
 
     display_name: Optional[str] = None
@@ -1606,7 +1606,7 @@ class UpdateToolResponseBodyJSONSchemaTool(BaseModel):
     json_schema: UpdateToolResponseBodyJSONSchema
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01M1XDKT9FH6JJYRTMP5SN4BTJ"
+        "tool_01M1XYQRHPBKSWD0WPXFZVXZE0"
     )
 
     display_name: Optional[str] = None
@@ -1804,7 +1804,7 @@ class UpdateToolResponseBodyFunctionTool(BaseModel):
     function: UpdateToolResponseBodyFunction
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "tool_01M1XDKT9FEDMF50QZ3J2JDVV0"
+        "tool_01M1XYQRHN0QC1BS6CQSRKXPK0"
     )
 
     display_name: Optional[str] = None